requirements.txt
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
idna==3.10
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
pyserial==3.5
requests==2.32.3
scp==0.15.0
urllib3==2.3.0
